requirements.txt
fastapi
loguru
matplotlib
notifiers
numpy
pydantic
pyserial
Requests
uvicorn
